refactor: move method to fix core dependencies
